templates/base/backend/python/requirements.txt
Flask
{{#if xsuaa}}
cfenv
sap-xssec
python-dotenv
{{else}}
{{/if}}